working on price graph

+++ jobs/price.job.js
@@ -25,7 +25,7 @@ function getPrice(conn_s, conn_i) {
     });
     response.on("end", function(err) {
       data = JSON.parse(buffer);
-      current_price = parseInt(data.query.results.quote.AskRealtime);
+      current_price = parseFloat(data.query.results.quote.AskRealtime);
       if (points.length === 0) {
         for (var i = 1; i <= 10; i++) {
           points.push({
